% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{address_fields_const}
\alias{address_fields_const}
\title{Construtor de especificação de colunas}
\usage{
address_fields_const(
  Address_or_Place = NULL,
  Address2 = NULL,
  Address3 = NULL,
  Neighborhood = NULL,
  City = NULL,
  County = NULL,
  State = NULL,
  ZIP = NULL,
  ZIP4 = NULL,
  Country = NULL
)
}
\arguments{
\item{Address_or_Place, Address2, Address3, Neighborhood, City, County, State, ZIP, ZIP4, Country}{Uma string. O nome da coluna que representa o respectivo campo de endereço
no dataframe de localizações. Pode ser \code{NULL}, no caso do campo não ser
especificado no dataframe. O valor de ao menos um dos campos não deve ser
nulo.}
}
\value{
Um vetor de caracteres com o nome da coluna que representa cada campo
no dataframe de localizações especificado.
}
\description{
Auxilia a criar um vetor de caracteres que especifica as colunas que
representam cada campo de endereço no dataframe de localizações.
}
