% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodepro.R
\docType{package}
\name{geocodepro}
\alias{geocodepro}
\alias{geocodepro-package}
\title{geocodepro: Geolocalização com o ArcGIS Pro}
\description{
Disponibiliza funções que facilitam a utilização da ferramenta de geolocalização do ArcGIS Pro no R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ipeaGIT/geocodepro}
  \item Report bugs at \url{https://github.com/ipeaGIT/geocodepro/issues}
}

}
\author{
\strong{Maintainer}: Daniel Herszenhut \email{dhersz@gmail.com} (\href{https://orcid.org/0000-0001-8066-1105}{ORCID})

}
\keyword{internal}
