% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geolocalização}
\usage{
geocode(
  locations,
  locator = "C://StreetMap/NewLocators/BRA/BRA.loc",
  address_fields = address_fields_const(),
  location_type = "ROUTING_LOCATION",
  output_fields = "MINIMAL",
  cache = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{locations}{Um dataframe. Os endereços que devem ser geolocalizados.
Deve conter todas as colunas especificadas em \code{address_fields}.}

\item{locator}{Uma string. O caminho até o localizador que deve ser
utilizado. Por padrão, utiliza o localizador incluído no ArcGIS Pro para o
Brasil.}

\item{address_fields}{Um vetor de caracteres. A correspondência entre as
colunas do dataframe e os campos de geolocalização utilizados pelo ArcGIS
Pro, em que os nomes do vetor são os campos esperados pela ferramenta e os
valores são as colunas correspondentes. A função \code{address_fields_const()}
serve para facilitar a criação deste vetor, fazendo também algumas
verificações em seu conteúdo. Note que itens com o valor \code{NULL} são
ignorados, e que a função deve receber ao menos um valor diferente de nulo.
Caso esteja criando o vetor manualmente, note que os nomes dos elementos
devem ser os mesmos nomes dos parâmetros da função
\code{address_fields_const()}.}

\item{location_type}{Uma string. Interfere na localização final de cada
endereço geolocalizado. As opções são \code{"ROUTING_LOCATION"} (padrão), em que
os endereços são posicionados ao longo da rua, e é considerado mais
apropriado para roteamento, ou \code{"ADDRESS_LOCATION"}, em que endereços são
localizados em uma posição equivalente ao "terraço" do endereço (no
centroide do polígono correspondente).}

\item{output_fields}{Uma string. As colunas que serão retornadas no output.
Possíveis opções são \code{"MINIMAL"} (padrão), que inclui as colunas do
dataframe de input e outros campos que descrevem o quão boa foi a
correspondência entre a geolocalização e o endereço, \code{"ALL"}, que inclui
todas as opções disponibilizadas pelo localizador, e \code{"LOCATION_ONLY"}, que
inclui apenas a geolocalização do endereço e as colunas do input.}

\item{cache}{Um logical (por padrão, \code{TRUE}) ou uma string. Se \code{TRUE}, é
utilizado um cache "geral" tanto para buscar localizações que já tenham
sido geolocalizadas antes quanto para adicionar quaisquer pontos que tenham
sido geolocalizados na operação corrente. Se \code{FALSE}, nenhum cache é usado.
Caso seja uma string, seu valor é utilizado para identificar um cache
específico. Por exemplo, pode ser útil usar um cache apenas para endereços
da RAIS, caso em que temos que passar \code{"rais"}. Caso desejemos invalidar
este cache e começar um novo cache da RAIS do zero, podemos passar
\code{"rais_v2"}.}

\item{verbose}{Um logical. Se mensagens informativas devem ser printadas
(padrão) ou não.}
}
\value{
Um objeto \code{sf} do tipo \code{POINT} com a localização dos endereços
especificados.
}
\description{
Geolocaliza um dataframe de endereços com base em diferentes atributos.
Documentação da ferramenta:
https://pro.arcgis.com/en/pro-app/latest/tool-reference/geocoding/geocode-addresses.htm.
}
\examples{
reticulate::use_condaenv(
  "C://Program Files/ArcGIS/Pro/bin/Python/envs/arcgispro-py3"
)

locations <- data.frame(
  logradouro = "Avenida Venceslau Brás, 72",
  bairro = "Botafogo",
  cidade = "Rio de Janeiro",
  uf = "RJ",
  cep = "22290-140"
)

result <- geocode(
  locations,
  address_fields = address_fields_const(
    Address_or_Place = "logradouro",
    Neighborhood = "bairro",
    City = "cidade",
    State = "uf",
    ZIP = "cep"
  ),
  cache = FALSE
)

}
